feat:add past-record page router and modify page file name
