+++ index.html
@@ -25,5 +25,10 @@
       crossorigin="anonymous"
     ></script>
     <script type="module" src="/src/main.js"></script>
+    <script
+      async
+      defer
+      src="https://cdnjs.cloudflare.com/ajax/libs/masonry/4.0.0/masonry.pkgd.min.js"
+    ></script>
   </body>
 </html>
