+++ src/assets/scss/layout/_past-record-page.scss
@@ -22,7 +22,7 @@
   @media (min-width: 768px) {
     flex-direction: column;
     align-self: flex-start;
-    width: 233px;
+    width: initial;
   }
   @media (min-width: 768px) {
     a + a {
@@ -30,31 +30,29 @@
     }
   }
 
-  .toggle-speech-record,
-  .toggle-past-works,
-  .toggle-related-news {
-    min-width: 98px;
-    min-height: 53px;
-    padding: 10px 15px;
-    font-size: 16px;
-    font-weight: 700;
-    color: #30363a;
-    background-color: #e3eaea;
-    border: 5px solid #a7ced4;
-    border-radius: 10px;
-    border-bottom-right-radius: 0;
-    &:hover {
-      color: #fff;
-      background-color: #a7ced4;
-      border: 5px solid #e3eaea;
-    }
-    @media (min-width: 768px) {
-      font-size: 28px;
-      font-weight: 700;
-      text-align: center;
-      width: 182px;
-      height: 80px;
-      line-height: 50px;
-    }
-  }
+  // .toggle-link {
+  //   min-width: 98px;
+  //   min-height: 53px;
+  //   padding: 10px 15px;
+  //   font-size: 16px;
+  //   font-weight: 700;
+  //   color: #30363a;
+  //   background-color: #e3eaea;
+  //   border: 5px solid #a7ced4;
+  //   border-radius: 10px;
+  //   border-bottom-right-radius: 0;
+  //   &:hover {
+  //     color: #fff;
+  //     background-color: #a7ced4;
+  //     border: 5px solid #e3eaea;
+  //   }
+  //   @media (min-width: 768px) {
+  //     font-size: 28px;
+  //     font-weight: 700;
+  //     text-align: center;
+  //     width: 182px;
+  //     height: 80px;
+  //     line-height: 50px;
+  //   }
+  // }
 }
