+++ src/views/RecordOfStreetMap.vue
