--- src/views/SpeechRecordPage.vue
@@ -1,239 +1,0 @@
-<script setup>
-import data from "../assets/JSON/data.json";
-const dataJson = data;
-
-function checkLength(array, index) {
-  if (array.length === 0) return "##";
-  if (array.length === 1 && index === 0) return array[0].link;
-  if (array.length === 1 && index === 1) return "##";
-  if (array.length === 2) return array[index].link;
-}
-
-// 佳憲重構1
-// function checkLength(array, index) {
-//   if (array.length === 0 || (array.length === 1 && index === 1)) return "##";
-//   else array.length === 2 || (array.length === 1 && index === 0);
-//   return array[index].link;
-// }
-
-// 佳憲重購2
-// function checkLength(array, index) {
-//   const apple = array.length;
-//   if (apple < 2 || apple === index) {
-//     return "##";
-//   } else {
-//     return array[index].link;
-//   }
-// }
-</script>
-
-<template>
-  <main>
-    <div class="past-record-container">
-      <div class="toggle-box">
-        <RouterLink class="toggle-speech-record" to="/SpeechRecordPage"
-          >演講紀錄</RouterLink
-        >
-
-        <RouterLink class="toggle-past-works" to="/PastWorksPage"
-          >過去作品</RouterLink
-        >
-
-        <RouterLink class="toggle-related-news" to="/RelatedNewsPage"
-          >相關新聞</RouterLink
-        >
-      </div>
-      <div class="sticky-notes">
-        <div class="speech-record">
-          <h2>演講紀錄</h2>
-          <div class="speech-record-topic">
-            <h3>活動主題</h3>
-            <h3>活動名稱</h3>
-            <h3>日期</h3>
-            <h3>相關資源</h3>
-          </div>
-          <div class="speech-record-container">
-            <div
-              class="speech-record-desktop"
-              v-for="record in dataJson.data"
-              :key="record.slides.link"
-            >
-              <div class="speech-record-content">
-                <p>{{ record.slides.title }}</p>
-                <p>{{ record.active }}</p>
-                <p>{{ record.date }}</p>
-                <div class="resources-content">
-                  <a target="_blank" :href="record.slides.link" class="tags"
-                    >投影片下載</a
-                  >
-                  <a
-                    target="_blank"
-                    :href="checkLength(record.ytVideos, 0)"
-                    class="tags"
-                    >影片連結</a
-                  >
-                  <a
-                    target="_blank"
-                    :href="checkLength(record.ytVideos, 1)"
-                    class="tags"
-                    >影片連結</a
-                  >
-                </div>
-              </div>
-            </div>
-            <div
-              class="speech-record-mobile"
-              v-for="record in dataJson.data"
-              :key="record.slides.link"
-            >
-              <div class="speech-record-topicNcontent">
-                <h3>活動主題</h3>
-                <p>{{ record.slides.title }}</p>
-              </div>
-              <div class="speech-record-topicNcontent">
-                <h3>活動名稱</h3>
-                <p>{{ record.active }}</p>
-              </div>
-              <div class="speech-record-topicNcontent">
-                <h3>日期</h3>
-                <p>{{ record.date }}</p>
-              </div>
-              <div class="speech-record-resources-topicNcontent">
-                <h3>相關資源</h3>
-                <div class="resources-content">
-                  <a target="_blank" :href="record.slides.link" class="tags"
-                    >投影片下載</a
-                  >
-                  <a
-                    target="_blank"
-                    :href="checkLength(record.ytVideos, 0)"
-                    class="tags"
-                    >影片連結</a
-                  >
-                  <a
-                    target="_blank"
-                    :href="checkLength(record.ytVideos, 1)"
-                    class="tags"
-                    >影片連結</a
-                  >
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </main>
-  <a href="#" class="top">Top &#8593;</a>
-</template>
-
-<style lang="scss" scoped>
-@import "../assets/scss/layout/past-record-page";
-@import "../assets/scss/components/sticky-notes";
-@import "../assets/scss/components/tags";
-
-.past-record-container .toggle-box .toggle-speech-record {
-  color: #fff;
-  background-color: #a7ced4;
-  border: 5px solid #e3eaea;
-}
-
-.speech-record {
-  padding: 14px 20px;
-
-  @media (min-width: 992px) {
-    padding: 24px 40px;
-  }
-
-  h2 {
-    font-weight: 700;
-    text-align: center;
-    margin-bottom: 33px;
-    font-size: 16px;
-
-    @media (min-width: 992px) {
-      font-size: 36px;
-    }
-  }
-  p {
-    font-size: 14px;
-    font-weight: 500;
-  }
-
-  // 針對past-speech-record
-
-  &-topic {
-    display: none;
-
-    @media (min-width: 992px) {
-      margin-top: 53px;
-      display: grid;
-      grid-template-columns: 2fr 2fr 1fr 3fr;
-      gap: 30px;
-
-      h3 {
-        font-size: 16px;
-        font-weight: 700;
-      }
-    }
-  }
-
-  &-desktop {
-    display: none;
-
-    @media (min-width: 992px) {
-      display: block;
-
-      .speech-record-content {
-        display: grid;
-        grid-template-columns: 2fr 2fr 1fr 3fr;
-        align-items: center;
-        border-bottom: white 1px solid;
-        padding: 20px 0;
-        gap: 30px;
-
-        .resources-content {
-          display: flex;
-          flex-direction: row;
-          justify-content: space-between;
-          gap: 10px;
-        }
-      }
-    }
-  }
-
-  &-mobile {
-    display: flex;
-    flex-direction: column;
-    gap: 12px;
-    border-bottom: #0000001a 1px solid;
-    padding: 20px 0;
-
-    @media (min-width: 992px) {
-      display: none;
-    }
-
-    h3 {
-      font-size: 14px;
-      font-weight: 700;
-    }
-
-    .speech-record-topicNcontent {
-      display: grid;
-      grid-template-columns: 1fr 3fr;
-    }
-
-    .speech-record-resources-topicNcontent {
-      display: flex;
-      flex-direction: column;
-      gap: 12px;
-
-      .resources-content {
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-      }
-    }
-  }
-}
-</style>
